pyproject.toml
[tool.poetry]
name = "dapla-toolbelt-pseudo"
version = "0.0.1"
description = "Pseudonymization extensions for Dapla Toolbelt"
authors = ["Team Skyinfrastruktur <dapla@ssb.no>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/statisticsnorway/dapla-toolbelt-pseudo"
repository = "https://github.com/statisticsnorway/dapla-toolbelt-pseudo"
documentation = "https://statisticsnorway.github.io/dapla-toolbelt-pseudo"
classifiers = [
    "Development Status :: 4 - Beta",
]
packages = [
    { include = "dapla_pseudo", from="src"},
]

[tool.poetry.urls]
Changelog = "https://github.com/statisticsnorway/dapla-toolbelt-pseudo/releases"

[tool.poetry.dependencies]
python = "^3.8"
dapla-toolbelt = "^1.3.6"
types-requests = "^2.28.11"
pydantic = "^1.10.2"
pyhumps = "^3.8.0"

[tool.poetry.dev-dependencies]
Pygments = ">=2.10.0"
black = ">=21.10b0"
coverage = {extras = ["toml"], version = ">=6.2"}
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bandit = ">=2.1.2"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = {version = ">=0.16.1"}
responses = "^0.22.0"
mock = "^4.0.3"
types-mock = "^4.0.15.2"

[tool.poetry.group.dev.dependencies]
responses = "^0.22.0"
devtools = "^0.9.0"
mock = "^4.0.3"

[tool.coverage.paths]
source = ["src/dapla_pseudo", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["src/dapla_pseudo"]

[tool.coverage.report]
show_missing = true
fail_under = 0

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.black]
line-length = 120


[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[[tool.mypy.overrides]]
module = [
    'dapla'
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
